requirements.txt
flask
requests
langchain-community
langchain-core
langchain-google-genai
langchain
dotenv
faiss-cpu
psycopg2-binary